Lots of bugfixes

+++ src/main/java/me/lauriichan/clay4j/IFont.java
@@ -4,6 +4,6 @@ public interface IFont {
     
     int id();
 
-    void calculateSize(String text, int fontSize, float[] size);
+    void calculateSize(String text, float fontSize, float[] size);
 
 }
